fixed weird issue where server was no longer returning lat/lon

+++ Server/Sources/Server/main.swift
@@ -76,8 +76,8 @@ func createRestaurants(request: HTTPRequest, response: HTTPResponse) {
         for json in requestJson {
             guard let name = json["name"] as? String,
                 let phoneNumber = json["phoneNumber"] as? String,
-                let latitude = json["latitude"] as? Double,
-                let longitude = json["longitude"] as? Double,
+                let latitude = json["latitude"] as? Float,
+                let longitude = json["longitude"] as? Float,
                 let website = json["website"] as? String else {
 
                     response.setBody(string: "Missing or Bad Parameter")
